+++ modules/apim-adapter/pom.xml
@@ -112,7 +112,6 @@
         <dependency>
             <groupId>com.jayway.jsonpath</groupId>
             <artifactId>json-path</artifactId>
-            <version>2.8.0</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
