+++ modules/apim-adapter/src/main/java/com/axway/apim/lib/utils/Utils.java
@@ -41,6 +41,7 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.ZoneId;
 import java.util.*;
+import java.util.stream.Collectors;
 
 public class Utils {
 
@@ -48,7 +49,7 @@ public class Utils {
     public static final String MASKED_VALUE = "********";
 
     public enum FedKeyType {
-        FilterCircuit("<key type='FilterCircuit'>"), OAuthAppProfile("<key type='OAuthAppProfile'>");
+        FilterCircuit("<key type='FilterCircuit'>"), OAuthAppProfile("<key type='OAuthAppProfile'>"), OAuthTokenProfile("<key type='AccessTokenPersist'>");
 
         private final String keyType;
 
@@ -468,5 +469,11 @@ public class Utils {
             }
             return "Err";
         }
+    }
+
+    public static Map<String,String> removeEmptyValuesFromMap(Map<String, String> map){
+        return map.entrySet().stream()
+            .filter(e -> !e.getValue().isEmpty())
+            .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue));
     }
 }
