+++ modules/apim-adapter/src/test/java/com/axway/apim/api/model/SecurityDeviceTest.java
@@ -1,45 +1,16 @@
 package com.axway.apim.api.model;
 
-import com.axway.apim.WiremockWrapper;
-import com.axway.apim.adapter.APIManagerAdapter;
-import com.axway.apim.lib.CoreParameters;
-import com.axway.apim.lib.error.AppException;
 import com.axway.apim.lib.utils.Utils;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.testng.Assert;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
 import java.util.List;
 
 public class SecurityDeviceTest {
-    //extends WiremockWrapper {
-//
-//    private APIManagerAdapter apiManagerAdapter;
-//
-//    @BeforeClass
-//    public void initWiremock() {
-//        super.initWiremock();
-//        CoreParameters coreParameters = new CoreParameters();
-//        coreParameters.setHostname("localhost");
-//        coreParameters.setUsername("test");
-//        coreParameters.setPassword(Utils.getEncryptedPassword());
-//        try {
-//            apiManagerAdapter = APIManagerAdapter.getInstance();
-//        } catch (AppException e) {
-//            throw new RuntimeException(e);
-//        }
-//    }
-//
-//    @AfterClass
-//    public void close() {
-//        Utils.deleteInstance(apiManagerAdapter);
-//
-//        super.close();
-//    }
+
 
     @Test
     public void compareSecurityDevice() throws JsonProcessingException {
