+++ modules/apim-adapter/src/test/java/com/axway/lib/utils/UtilsTest.java
@@ -367,5 +367,4 @@ public class UtilsTest {
             Assert.assertEquals(99, Utils.handleAppException(exception, LoggerFactory.getLogger("Utils"), new ErrorCodeMapper()));
         }
     }
-
 }
