+++ modules/apis/src/main/java/com/axway/apim/apiimport/APIImportManager.java
@@ -53,7 +53,7 @@ public class APIImportManager {
         } else {
             if (!changeState.hasAnyChanges()) {
                 APIPropertiesExport.getInstance().setProperty("feApiId", changeState.getActualAPI().getId());
-                LOG.debug("BUT, no changes detected between Import- and API-Manager-API. Exiting now...");
+                LOG.debug("No changes detected between Import- and API-Manager-API. Exiting now...");
                 throw new AppException("No changes detected between Import- and API-Manager-API: '" + changeState.getActualAPI().getName() + "' (" + changeState.getActualAPI().getId() + ")", ErrorCode.NO_CHANGE);
             }
             LOG.info("Recognized the following changes. Potentially Breaking: {} plus Non-Breaking: {}", changeState.getBreakingChanges(), changeState.getNonBreakingChanges());
