+++ modules/spectoconfig/pom.xml
@@ -32,7 +32,6 @@
         <dependency>
             <groupId>com.jayway.jsonpath</groupId>
             <artifactId>json-path</artifactId>
-            <version>2.8.0</version>
             <scope>test</scope>
         </dependency>
         <dependency>
